+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to `libcasm-clexulator` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0a4] - 2023-10-26
+## [2.0a4] - Unreleased
 
 ### Added
 
 - Added gradients of correlations functionality in C++, Python functions
 - Added FADBAD library to support compiling of Clexulators capable of calculating gradients 
+- Added methods to libcasm.clexulator.PrimNeighborList to construct a PrimNeighborList with default parameters for a libcasm.xtal.Prim and to expand and access the neighbors.
+- Added methods to libcasm.clexulator.SuperNeighborList to access the neighbor sites and unitcells.
+
 
 ## [2.0a3] - 2023-10-25